Skip insignificant bytes in RadixSorter for speed
This commit also updates the tests to reflect this behavior.
In particular, the input data to both TimSorter and RadixSorter
are modified in such a way that it doesn't matter whether these
bytes are compared or not. This ensures correctness with our
sorter even with ignored bytes.

Result: the time it takes to sort 200m Long's went down from 24.1s
to 17.2s. These are medians over 10 iterations. The microbenchmark
code will be added in the subsequent commit.