Guard against fitting failures + fix loss indices
The solution now is to retry once with starting parameters equal
to the average of all the parameters fitted so far, then if that
fails retry again with decay = 1.

This commit also fixes the issue where the loss indices (x in
the prediction) always starts at 0. This is no longer the case.